use gpg explicit in message_view()
All messages now store if they need gpg interaction to be usable.
The functions CMessage:need_gpg() in C++ and Message:need_gpg() in Lua
can be used to retrieve this information.

message_view() replaces the content of the opened message with a temporary
file returned by gpg.decrypt() if needed.

The message flags in index_mode now show a 'G' if the message is either
signed or encrypted instead of 'S'.